parser: split into pat.rs
